+++ src/app/dataset-flow/flows-table/flows-table.helpers.spec.ts
@@ -69,7 +69,7 @@ describe("FlowTableHelpers", () => {
                 watermarkModified: true,
                 forBreakingChange: FlowTriggerBreakingChangeRule.Recover,
             }),
-        ).toEqual("waiting for a reactive condition");
+        ).toEqual("waiting for input data: 100/500");
     });
 
     it("should check waiting block text with FlowStartConditionExecutor typename", () => {
@@ -184,10 +184,26 @@ describe("FlowTableHelpers", () => {
         expect(FlowTableHelpers.descriptionSubMessage(mockTableFlowSummaryDataFragments[6])).toContain("Aborted at");
     });
 
-    it(`should check description end of message with description FlowDescriptionDatasetPollingIngest typename and failed outcome `, () => {
+    it(`should check description end of message with description FlowDescriptionDatasetPollingIngest typename and failed recoverable outcome `, () => {
         expect(FlowTableHelpers.descriptionSubMessage(mockTableFlowSummaryDataFragments[7])).toEqual(
             "An error occurred, see logs for more details",
         );
+    });
+
+    it(`should check description end of message with description FlowDescriptionDatasetPollingIngest typename and failed unrecoverable outcome `, () => {
+        expect(
+            FlowTableHelpers.descriptionSubMessage({
+                ...mockTableFlowSummaryDataFragments[7],
+                outcome: {
+                    __typename: "FlowFailedError",
+                    reason: {
+                        __typename: "TaskFailureReasonGeneral",
+                        message: "Failed",
+                        recoverable: false,
+                    },
+                },
+            }),
+        ).toEqual("An unrecoverable error occurred, see logs for more details");
     });
 
     it(`should check description end of message with description FlowDescriptionDatasetPollingIngest typename and cacheable source`, () => {