+++ src/app/dataset-flow/flows-table/flows-table.helpers.ts
@@ -10,6 +10,7 @@ import {
     FlowStatus,
     FlowSummaryDataFragment,
     FlowTimingRecords,
+    FlowTriggerBatchingRuleBuffering,
 } from "src/app/api/kamu.graphql.interface";
 import { MaybeNull } from "src/app/interface/app.types";
 import AppValues from "src/app/common/values/app.values";
@@ -208,7 +209,7 @@ export class FlowTableHelpers {
                     case "FlowFailedError": {
                         switch (element.outcome.reason.__typename) {
                             case "TaskFailureReasonGeneral":
-                                return `An error occurred, see logs for more details`;
+                                return `An${element.outcome.reason.recoverable ? "" : " unrecoverable"} error occurred, see logs for more details`;
                             case "TaskFailureReasonInputDatasetCompacted": {
                                 return `Input dataset <a class="text-small text-danger">${element.outcome.reason.inputDataset.name}</a> was hard compacted`;
                             }
@@ -368,7 +369,7 @@ export class FlowTableHelpers {
                 return "waiting for a throttling condition";
 
             case "FlowStartConditionReactive":
-                return "waiting for a reactive condition";
+                return `waiting for input data: ${startCondition.accumulatedRecordsCount}/${(startCondition.activeBatchingRule as FlowTriggerBatchingRuleBuffering).minRecordsToAwait}`;
 
             case "FlowStartConditionExecutor": {
                 return "waiting for a free executor";