+++ src/app/dataset-view/additional-components/dataset-settings-component/dataset-settings.model.ts
@@ -26,6 +26,11 @@ export enum BatchingRuleType {
     BUFFERING = "Buffering",
 }
 
+export enum FlowTriggerStopPolicyType {
+    NEVER = "Never",
+    AFTER_CONSECUTIVE_FAILURES = "AfterConsecutiveFailures",
+}
+
 export interface DatasetSettingsSidePanelItem {
     name: string;
     iconName: string;