+++ src/app/dataset-view/additional-components/dataset-settings-component/tabs/scheduling/dataset-settings-scheduling-tab.component.ts
@@ -6,9 +6,9 @@
  */
 
 import { AfterViewInit, ChangeDetectionStrategy, ChangeDetectorRef, Component, inject, Input } from "@angular/core";
-import { FormGroup, FormsModule, ReactiveFormsModule } from "@angular/forms";
+import { FormControl, FormGroup, FormsModule, ReactiveFormsModule } from "@angular/forms";
 import { BaseComponent } from "src/app/common/components/base.component";
-import { ScheduleType } from "../../dataset-settings.model";
+import { ScheduleType, FlowTriggerStopPolicyType } from "../../dataset-settings.model";
 import { DatasetBasicsFragment, DatasetFlowType, DatasetKind } from "src/app/api/kamu.graphql.interface";
 import { DatasetFlowTriggerService } from "../../services/dataset-flow-trigger.service";
 import { takeUntilDestroyed } from "@angular/core/rxjs-interop";
@@ -18,9 +18,13 @@ import { NgIf } from "@angular/common";
 import { MatDividerModule } from "@angular/material/divider";
 import { MaybeNull } from "src/app/interface/app.types";
 import { IngestTriggerFormValue } from "./ingest-trigger-form/ingest-trigger-form.types";
-import { MatProgressBarModule } from "@angular/material/progress-bar";
 import { DatasetSettingsSchedulingTabData } from "./dataset-settings-scheduling-tab.data";
 import { SchedulingSettingsFormType, SchedulingSettingsFormValue } from "./dataset-settings-scheduling-tab.types";
+import { FlowStopPolicyFormComponent } from "../shared/flow-stop-policy-form/flow-stop-policy-form.component";
+import { FlowStopPolicyFormValue } from "../shared/flow-stop-policy-form/flow-stop-policy-form.types";
+import { FlowTooltipsTexts } from "src/app/common/tooltips/flow-tooltips.text";
+import { TooltipIconComponent } from "src/app/common/components/tooltip-icon/tooltip-icon.component";
+import { MatSlideToggleModule } from "@angular/material/slide-toggle";
 
 @Component({
     selector: "app-dataset-settings-scheduling-tab",
@@ -35,11 +39,13 @@ import { SchedulingSettingsFormType, SchedulingSettingsFormValue } from "./datas
         ReactiveFormsModule,
 
         //-----//
-        MatProgressBarModule,
+        MatSlideToggleModule,
         MatDividerModule,
 
         //-----//
         IngestTriggerFormComponent,
+        FlowStopPolicyFormComponent,
+        TooltipIconComponent,
     ],
 })
 export class DatasetSettingsSchedulingTabComponent extends BaseComponent implements AfterViewInit {
@@ -49,8 +55,12 @@ export class DatasetSettingsSchedulingTabComponent extends BaseComponent impleme
     private readonly cdr = inject(ChangeDetectorRef);
     private readonly datasetFlowTriggerService = inject(DatasetFlowTriggerService);
 
+    public readonly UPDATES_TOOLTIP = FlowTooltipsTexts.UPDATE_SELECTOR_TOOLTIP;
+
     public readonly form: FormGroup<SchedulingSettingsFormType> = new FormGroup<SchedulingSettingsFormType>({
+        updatesEnabled: new FormControl<boolean>({ value: false, disabled: false }, { nonNullable: true }),
         ingestTrigger: IngestTriggerFormComponent.buildForm(),
+        stopPolicy: FlowStopPolicyFormComponent.buildForm(),
     });
 
     public get datasetBasics(): DatasetBasicsFragment {
@@ -64,7 +74,9 @@ export class DatasetSettingsSchedulingTabComponent extends BaseComponent impleme
     public ngAfterViewInit() {
         this.form.patchValue(
             {
+                updatesEnabled: this.schedulingTabData.paused === false,
                 ingestTrigger: this.buildInitialIngestTriggerFormValue(),
+                stopPolicy: this.buildInitialStopPolicyFormValue(),
             } as SchedulingSettingsFormValue,
             { emitEvent: true },
         );
@@ -78,11 +90,9 @@ export class DatasetSettingsSchedulingTabComponent extends BaseComponent impleme
             return null;
         }
 
-        const updatesEnabled = !this.schedulingTabData.paused;
         switch (schedule.__typename) {
             case "TimeDelta": {
                 return {
-                    updatesEnabled,
                     __typename: ScheduleType.TIME_DELTA,
                     timeDelta: {
                         every: schedule.every,
@@ -96,7 +106,6 @@ export class DatasetSettingsSchedulingTabComponent extends BaseComponent impleme
 
             case "Cron5ComponentExpression": {
                 return {
-                    updatesEnabled,
                     __typename: ScheduleType.CRON_5_COMPONENT_EXPRESSION,
                     timeDelta: {
                         every: null,
@@ -115,22 +124,58 @@ export class DatasetSettingsSchedulingTabComponent extends BaseComponent impleme
         }
     }
 
-    public saveScheduledUpdates(): void {
+    private buildInitialStopPolicyFormValue(): MaybeNull<FlowStopPolicyFormValue> {
+        const stopPolicy = this.schedulingTabData.stopPolicy;
+        if (!stopPolicy) {
+            return null;
+        }
+
+        switch (stopPolicy.__typename) {
+            case "FlowTriggerStopPolicyNever":
+                return {
+                    stopPolicyType: FlowTriggerStopPolicyType.NEVER,
+                    maxFailures: FlowStopPolicyFormComponent.DEFAULT_MAX_FAILURES,
+                };
+
+            case "FlowTriggerStopPolicyAfterConsecutiveFailures":
+                return {
+                    stopPolicyType: FlowTriggerStopPolicyType.AFTER_CONSECUTIVE_FAILURES,
+                    maxFailures: stopPolicy.maxFailures,
+                };
+
+            /* istanbul ignore next */
+            default: {
+                throw new Error(`Unknown stop policy type: ${stopPolicy.__typename}`);
+            }
+        }
+    }
+
+    public saveUpdates(): void {
         const formValue = this.form.getRawValue() as SchedulingSettingsFormValue;
         const ingestTriggerFormValue = formValue.ingestTrigger;
 
-        this.datasetFlowTriggerService
-            .setDatasetFlowTriggers({
-                datasetId: this.datasetBasics.id,
-                datasetFlowType: DatasetFlowType.Ingest,
-                paused: !ingestTriggerFormValue.updatesEnabled,
-                triggerInput: IngestTriggerFormComponent.buildPollingTriggerInput(ingestTriggerFormValue),
-                datasetInfo: {
-                    accountName: this.datasetBasics.owner.accountName,
-                    datasetName: this.datasetBasics.name,
-                },
-            })
-            .pipe(takeUntilDestroyed(this.destroyRef))
-            .subscribe();
+        if (formValue.updatesEnabled) {
+            this.datasetFlowTriggerService
+                .setDatasetFlowTrigger({
+                    datasetId: this.datasetBasics.id,
+                    datasetFlowType: DatasetFlowType.Ingest,
+                    triggerRuleInput: IngestTriggerFormComponent.buildPollingTriggerRuleInput(ingestTriggerFormValue),
+                    triggerStopPolicyInput: FlowStopPolicyFormComponent.buildStopPolicyInput(formValue.stopPolicy),
+                    datasetInfo: {
+                        accountName: this.datasetBasics.owner.accountName,
+                        datasetName: this.datasetBasics.name,
+                    },
+                })
+                .pipe(takeUntilDestroyed(this.destroyRef))
+                .subscribe();
+        } else {
+            this.datasetFlowTriggerService
+                .pauseDatasetFlowTrigger({
+                    datasetId: this.datasetBasics.id,
+                    datasetFlowType: DatasetFlowType.Ingest,
+                })
+                .pipe(takeUntilDestroyed(this.destroyRef))
+                .subscribe();
+        }
     }
 }