+++ src/app/dataset-view/additional-components/dataset-settings-component/tabs/scheduling/dataset-settings-scheduling-tab.types.ts
@@ -5,13 +5,21 @@
  * included in the LICENSE file.
  */
 
-import { FormGroup } from "@angular/forms";
+import { FormControl, FormGroup } from "@angular/forms";
 import { IngestTriggerFormType, IngestTriggerFormValue } from "./ingest-trigger-form/ingest-trigger-form.types";
+import {
+    FlowStopPolicyFormType,
+    FlowStopPolicyFormValue,
+} from "../shared/flow-stop-policy-form/flow-stop-policy-form.types";
 
 export interface SchedulingSettingsFormType {
+    updatesEnabled: FormControl<boolean>;
     ingestTrigger: FormGroup<IngestTriggerFormType>;
+    stopPolicy: FormGroup<FlowStopPolicyFormType>;
 }
 
 export interface SchedulingSettingsFormValue {
+    updatesEnabled: boolean;
     ingestTrigger: IngestTriggerFormValue;
+    stopPolicy: FlowStopPolicyFormValue;
 }