+++ src/app/dataset-view/additional-components/dataset-settings-component/tabs/transform-options/dataset-settings-transform-options-tab.component.ts
@@ -7,7 +7,7 @@
 
 import { AfterViewInit, ChangeDetectionStrategy, ChangeDetectorRef, Component, inject, Input } from "@angular/core";
 import { takeUntilDestroyed } from "@angular/core/rxjs-interop";
-import { FormGroup, FormsModule, ReactiveFormsModule } from "@angular/forms";
+import { FormControl, FormGroup, FormsModule, ReactiveFormsModule } from "@angular/forms";
 import { DatasetBasicsFragment, DatasetFlowType, DatasetKind, TimeUnit } from "src/app/api/kamu.graphql.interface";
 import { DatasetFlowTriggerService } from "../../services/dataset-flow-trigger.service";
 import { BaseComponent } from "src/app/common/components/base.component";
@@ -15,14 +15,16 @@ import RoutingResolvers from "src/app/common/resolvers/routing-resolvers";
 import { MatDividerModule } from "@angular/material/divider";
 import { TransformTriggerFormComponent } from "./transform-trigger-form/transform-trigger-form.component";
 import { NgIf } from "@angular/common";
-import {
-    TransformSettingsFormType,
-    TransformSettingsFormValue,
-} from "./dataset-settings-transform-options-tab.component.types";
+import { TransformSettingsFormType, TransformSettingsFormValue } from "./dataset-settings-transform-options-tab.types";
 import { MaybeNull } from "src/app/interface/app.types";
 import { TransformTriggerFormValue } from "./transform-trigger-form/transform-trigger-form.types";
 import { DatasetSettingsTransformOptionsTabData } from "./dataset-settings-transform-options-tab.data";
-import { BatchingRuleType } from "../../dataset-settings.model";
+import { BatchingRuleType, FlowTriggerStopPolicyType } from "../../dataset-settings.model";
+import { FlowStopPolicyFormComponent } from "../shared/flow-stop-policy-form/flow-stop-policy-form.component";
+import { FlowTooltipsTexts } from "src/app/common/tooltips/flow-tooltips.text";
+import { TooltipIconComponent } from "src/app/common/components/tooltip-icon/tooltip-icon.component";
+import { MatSlideToggleModule } from "@angular/material/slide-toggle";
+import { FlowStopPolicyFormValue } from "../shared/flow-stop-policy-form/flow-stop-policy-form.types";
 
 @Component({
     selector: "app-dataset-settings-transform-options-tab",
@@ -37,10 +39,13 @@ import { BatchingRuleType } from "../../dataset-settings.model";
         ReactiveFormsModule,
 
         //-----//
+        MatSlideToggleModule,
         MatDividerModule,
 
         //-----//
         TransformTriggerFormComponent,
+        FlowStopPolicyFormComponent,
+        TooltipIconComponent,
     ],
 })
 export class DatasetSettingsTransformOptionsTabComponent extends BaseComponent implements AfterViewInit {
@@ -50,6 +55,8 @@ export class DatasetSettingsTransformOptionsTabComponent extends BaseComponent i
     private readonly cdr = inject(ChangeDetectorRef);
     private readonly datasetFlowTriggerService = inject(DatasetFlowTriggerService);
 
+    public readonly UPDATES_TOOLTIP = FlowTooltipsTexts.UPDATE_SELECTOR_TOOLTIP;
+
     public get datasetBasics(): DatasetBasicsFragment {
         return this.transformTabData.datasetBasics;
     }
@@ -57,14 +64,19 @@ export class DatasetSettingsTransformOptionsTabComponent extends BaseComponent i
     public get isDerivedDataset(): boolean {
         return this.datasetBasics.kind === DatasetKind.Derivative;
     }
+
     public readonly form: FormGroup<TransformSettingsFormType> = new FormGroup<TransformSettingsFormType>({
+        updatesEnabled: new FormControl<boolean>({ value: false, disabled: false }, { nonNullable: true }),
         transformTrigger: TransformTriggerFormComponent.buildForm(),
+        stopPolicy: FlowStopPolicyFormComponent.buildForm(),
     });
 
     public ngAfterViewInit() {
         this.form.patchValue(
             {
+                updatesEnabled: this.transformTabData.paused === false,
                 transformTrigger: this.buildInitialTransformTriggerFormValue(),
+                stopPolicy: this.buildInitialStopPolicyFormValue(),
             } as TransformSettingsFormValue,
             { emitEvent: true },
         );
@@ -78,7 +90,6 @@ export class DatasetSettingsTransformOptionsTabComponent extends BaseComponent i
             return null;
         }
 
-        const updatesEnabled = !this.transformTabData.paused;
         switch (reactive.forNewData.__typename) {
             case "FlowTriggerBatchingRuleBuffering": {
                 // Since we are disabling Days and Weeks in batching rule UI, check for special values
@@ -98,7 +109,6 @@ export class DatasetSettingsTransformOptionsTabComponent extends BaseComponent i
                 }
 
                 return {
-                    updatesEnabled,
                     forNewData: {
                         batchingRuleType: BatchingRuleType.BUFFERING,
                         buffering: {
@@ -115,7 +125,6 @@ export class DatasetSettingsTransformOptionsTabComponent extends BaseComponent i
 
             case "FlowTriggerBatchingRuleImmediate":
                 return {
-                    updatesEnabled,
                     forNewData: {
                         batchingRuleType: BatchingRuleType.IMMEDIATE,
                     },
@@ -127,22 +136,61 @@ export class DatasetSettingsTransformOptionsTabComponent extends BaseComponent i
         }
     }
 
+    private buildInitialStopPolicyFormValue(): MaybeNull<FlowStopPolicyFormValue> {
+        const stopPolicy = this.transformTabData.stopPolicy;
+        if (!stopPolicy) {
+            return null;
+        }
+
+        switch (stopPolicy.__typename) {
+            case "FlowTriggerStopPolicyNever":
+                return {
+                    stopPolicyType: FlowTriggerStopPolicyType.NEVER,
+                    maxFailures: FlowStopPolicyFormComponent.DEFAULT_MAX_FAILURES,
+                };
+
+            case "FlowTriggerStopPolicyAfterConsecutiveFailures":
+                return {
+                    stopPolicyType: FlowTriggerStopPolicyType.AFTER_CONSECUTIVE_FAILURES,
+                    maxFailures: stopPolicy.maxFailures,
+                };
+
+            /* istanbul ignore next */
+            default: {
+                throw new Error(`Unknown stop policy type: ${stopPolicy.__typename}`);
+            }
+        }
+    }
+
     public saveUpdates(): void {
         const formValue = this.form.getRawValue() as TransformSettingsFormValue;
-        const transformTriggerFormValue = formValue.transformTrigger;
+        const updatesEnabled = formValue.updatesEnabled;
 
-        this.datasetFlowTriggerService
-            .setDatasetFlowTriggers({
-                datasetId: this.datasetBasics.id,
-                datasetFlowType: DatasetFlowType.ExecuteTransform,
-                paused: !transformTriggerFormValue.updatesEnabled,
-                triggerInput: TransformTriggerFormComponent.buildTransformTriggerInput(transformTriggerFormValue),
-                datasetInfo: {
-                    accountName: this.datasetBasics.owner.accountName,
-                    datasetName: this.datasetBasics.name,
-                },
-            })
-            .pipe(takeUntilDestroyed(this.destroyRef))
-            .subscribe();
+        if (updatesEnabled) {
+            const transformTriggerFormValue = formValue.transformTrigger;
+
+            this.datasetFlowTriggerService
+                .setDatasetFlowTrigger({
+                    datasetId: this.datasetBasics.id,
+                    datasetFlowType: DatasetFlowType.ExecuteTransform,
+                    triggerRuleInput:
+                        TransformTriggerFormComponent.buildTransformTriggerRuleInput(transformTriggerFormValue),
+                    triggerStopPolicyInput: FlowStopPolicyFormComponent.buildStopPolicyInput(formValue.stopPolicy),
+                    datasetInfo: {
+                        accountName: this.datasetBasics.owner.accountName,
+                        datasetName: this.datasetBasics.name,
+                    },
+                })
+                .pipe(takeUntilDestroyed(this.destroyRef))
+                .subscribe();
+        } else {
+            this.datasetFlowTriggerService
+                .pauseDatasetFlowTrigger({
+                    datasetId: this.datasetBasics.id,
+                    datasetFlowType: DatasetFlowType.ExecuteTransform,
+                })
+                .pipe(takeUntilDestroyed(this.destroyRef))
+                .subscribe();
+        }
     }
 }